TRT-2049: Move watch counts test
